// src/app/models/route-instance.ts
export class RouteInstance {
    id: number = undefined;
    route_id: number = undefined;
    leader_name: string = undefined;
    leader_phone: string = undefined;
    scribe_name: string = undefined;
    scribe_phone: string = undefined;
    heat_route: boolean = undefined;
    number_route_members: number = undefined;
    start_time: Date = undefined;
    end_time: Date = undefined;
  }